# backend/routes/meetings.py
# backend/routers/meetings.py

# Copyright 2025 Alun King
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

from fastapi import APIRouter, Depends, HTTPException, Query, Header
from sqlalchemy import and_
from sqlalchemy.orm import Session
from backend.models import Meeting, MeetingOut, MeetingComment, MeetingCommentOut, GroupMember, GroupMemberOut, RawFile, Group
from backend.db_dependency import get_db
from datetime import date, datetime, timedelta
from backend.validation import MeetingCreateEdit, MeetingCommentCreate, MeetingAttendee
from backend.jobs.service import submit_job
from backend.llm.ollama_client import OllamaError
from backend.summarization.summariser import generate_meeting_summary

router = APIRouter(prefix="/groups/{group_id}/meetings", tags=["meetings"])
from backend.auth import (
    EmailPrincipal,
    is_email_workflow_group_member,
    is_group_member,
    is_group_owner,
    is_group_owner_or_email_member,
)

@router.post("/")
def create_meeting(
        group_id: int,
        meeting_data:MeetingCreateEdit,
        db: Session = Depends(get_db), 
        _principal: EmailPrincipal = Depends(is_email_workflow_group_member),
        idempotency_key: str | None = Header(default=None, alias="Idempotency-Key"),
    ):
    if idempotency_key:
        existing = db.query(Meeting).filter(Meeting.idempotency_key == idempotency_key).first()
        if existing is not None:
            if existing.group_id != group_id:
                raise HTTPException(status_code=409, detail="Idempotency key belongs to another group")
            return existing

    new_meeting = Meeting(group_id=group_id, date=meeting_data.date, created=datetime.now())
    new_meeting.idempotency_key = idempotency_key
    
    db.add(new_meeting)
    db.commit()
    db.refresh(new_meeting)
    return new_meeting

@router.get("/")
def list_meetings(
        group_id: int,
        from_date: date | None = Query(None, description="Include meetings on or after this date"),
        to_date: date | None = Query(None, description="Include meetings on or before this date"),
        db: Session = Depends(get_db), 
        user_id: int = Depends(is_group_member)
    ):
    if from_date and to_date and from_date > to_date:
        raise HTTPException(status_code=400, detail="from_date must be on or before to_date")

    query = db.query(Meeting).filter(Meeting.group_id == group_id)
    if from_date:
        query = query.filter(Meeting.date >= datetime.combine(from_date, datetime.min.time()))
    if to_date:
        exclusive_end = datetime.combine(to_date + timedelta(days=1), datetime.min.time())
        query = query.filter(Meeting.date < exclusive_end)
    return query.order_by(Meeting.date).all()

@router.get("/{meeting_id}", response_model=MeetingOut)
def get_meeting(
        group_id: int,
        meeting_id: int, 
        db: Session = Depends(get_db), 
        user_id: int = Depends(is_group_member)
    ):
    meeting = db.query(Meeting).filter(
        and_(Meeting.id == meeting_id, Meeting.group_id == group_id)
    ).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")
    return meeting

@router.post("/{meeting_id}/comments", response_model=MeetingCommentOut, status_code=201)
def add_meeting_comment(
        group_id: int,
        meeting_id: int,
        comment_data: MeetingCommentCreate,
        db: Session = Depends(get_db),
        principal: EmailPrincipal = Depends(is_email_workflow_group_member)
    ):
    meeting = db.query(Meeting).filter(
        and_(Meeting.id == meeting_id, Meeting.group_id == group_id)
    ).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")

    comment = MeetingComment(
        meeting_id=meeting_id,
        user_id=principal.user_id,
        group_member_id=principal.scoped_group_member_id,
        comment=comment_data.comment,
    )
    db.add(comment)
    db.commit()
    db.refresh(comment)
    return comment

@router.delete("/{meeting_id}")
def delete_meeting(
        group_id: int,
        meeting_id: int,
        db: Session = Depends(get_db),
        user_id: int = Depends(is_group_owner)
    ):
    meeting = db.query(Meeting).filter(Meeting.id == meeting_id).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")
    db.delete(meeting)
    db.commit()
    return {"message": "Meeting deleted"}

@router.post("/{meeting_id}/attendees", response_model=GroupMemberOut)
def add_attendee(
        group_id:int,
        meeting_id:int,
        attendee_data:MeetingAttendee,
        db:Session = Depends(get_db),
        principal: EmailPrincipal = Depends(is_group_owner_or_email_member)
    ):
    print("Recieved attendee data:", attendee_data)
    if principal.scoped_group_member_id is not None and not attendee_data.member_id:
        raise HTTPException(status_code=403, detail="Group members can only attach known attendees")
    meeting = db.query(Meeting).filter(
        and_(Meeting.id == meeting_id, Meeting.group_id == group_id)
    ).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")

    #if the member exists, then add them to the attendance.
    if attendee_data.member_id:
        # Add existing member to the meeting
        member = db.query(GroupMember).filter(GroupMember.id == attendee_data.member_id).first()
        if not member or group_id not in [group.id for group in member.groups]:
            raise HTTPException(status_code=404, detail="Member not found")

    elif attendee_data.guest and attendee_data.name:
        # Create a new guest group member
        member = GroupMember(name=attendee_data.name)
        db.add(member)
        db.commit()
        db.refresh(member)
    else:
        raise HTTPException(status_code=400, detail="Invalid attendee data")

    # Link the member to the meeting
    if member not in meeting.attendees:
        meeting.attendees.append(member)
        db.commit()

    return member

@router.delete("/{meeting_id}/attendees/{member_id}")
def remove_attendee(
    group_id: int,
    meeting_id: int,
    member_id: int,
    db: Session = Depends(get_db),
    user_id: int = Depends(is_group_owner)
):
    # Check if meeting exists
    meeting = db.query(Meeting).filter(Meeting.id == meeting_id).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")

    # Check if member exists
    member = db.query(GroupMember).filter(GroupMember.id == member_id).first()
    if not member:
        raise HTTPException(status_code=404, detail="Member not found")

    # Remove the member from the meeting attendees
    if member in meeting.attendees:
        meeting.attendees.remove(member)
        db.commit()
        return {"message": "Attendee removed"}
    else:
        raise HTTPException(status_code=404, detail="Member is not an attendee of this meeting")
    
@router.post("/{meeting_id}/transcribe")
async def start_transcription_job(
    group_id: int,
    meeting_id: int,
    reprocess: bool = Query(False),
    db: Session = Depends(get_db),
    user_id: int = Depends(is_group_owner),
):
    meeting = db.query(Meeting).filter(
        and_(Meeting.id == meeting_id, Meeting.group_id == group_id)
    ).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")

    # Find first audio file for this meeting
    audio_file = db.query(RawFile).filter(
        and_(
            RawFile.meeting_id == meeting_id,
            RawFile.type == "audio"
        )
    ).first()

    if not audio_file:
        raise HTTPException(
            status_code=400,
            detail="No audio files found for this meeting. Please upload an audio file before starting transcription."
        )

    # Prevent accidental reprocessing
    if audio_file.processed_date is not None and not reprocess:
        raise HTTPException(
            status_code=409,
            detail="This file has already been processed. To reprocess, set the 'reprocess=true' query parameter."
        )

    job = submit_job(
        db, "transcription", {"group_id": group_id, "meeting_id": meeting_id},
        user_id=user_id, group_id=group_id, meeting_id=meeting_id,
    )

    return {
        "message": f"{'Reprocessing' if reprocess else 'Transcription job started'}.",
        "file": audio_file.file_name,
        "job_id": job.id,
        "status_check_url": f"/jobs/{job.id}",
    }


@router.get("/{meeting_id}/summarise")
def summarise_meeting(
        group_id: int,
        meeting_id: int,
        regenerate: bool = Query(False),
        db: Session = Depends(get_db),
        user_id: int = Depends(is_group_owner)
    ):
    """Returns this meeting's summary, generating it via the local LLM
    (backend/summarization) on first request if none is cached yet. Pass
    ?regenerate=true to force a fresh summary even if one is already stored -
    e.g. after the transcript has been corrected."""
    meeting = db.query(Meeting).filter(
        and_(Meeting.id == meeting_id, Meeting.group_id == group_id)
    ).first()
    if not meeting:
        raise HTTPException(status_code=404, detail="Meeting not found")

    if meeting.summary and not regenerate:
        return {
            "summary": meeting.summary,
            "generated_at": meeting.summary_generated_at,
            "cached": True,
        }

    group = db.query(Group).get(group_id)
    try:
        summary = generate_meeting_summary(db, group, meeting)
    except FileNotFoundError:
        raise HTTPException(status_code=404, detail="No transcript available for this meeting yet")
    except OllamaError as exc:
        raise HTTPException(status_code=502, detail=str(exc))

    return {
        "summary": summary,
        "generated_at": meeting.summary_generated_at,
        "cached": False,
    }